requirements.txt
google-genai
python-dotenv
uvicorn
fastapi